cardio cleanup
