+++ card/lib/card/cache.rb
@@ -118,7 +118,7 @@ class Card
       # (the non-standard caches)
       def reset_other
         Card::Codename.reset_cache
-        Cardio.delete_tmp_files
+        Card.delete_tmp_files
       end
 
       # generate a cache key from an object
